멤버 id, 그룹 id를 slackClient 로 이동
